Fixes for footer layout in IE11.
There are some issues with flexbox in IE11, this fixes our current
rendering problem where the footer would float above the site content.

This fix changes how things are rendered in IE. In Safari/Chrome/Edge it
renders as intended with the footer pinned to the bottom of the screen.
In IE11 the footer renders as a regular div, just a block attached to
the bottom of the main content.

<main> is not a supported tag in IE11, it renders fine but errors in the
console so I'm happy to put it back in if folks don't mind the error.