Remove S3ID from Upload.
We had originally intended for an upload's S3ID to allow us to keep our internal
database identifiers out of S3. However, since we're using an upload's document and
move to namespace object keys, we are using database ids to construct S3 keys already.

For the sake of simplicity, we're removing the S3ID from upload and accepting that our S3
keys will be tightly coupled to our database identifiers.