the /regex literal/ syntax seems to make Fields behave badly.
For an email Field, passing a regex literal here instead of a string caused a very strange bug
I still don't fully understand. The component stopped updating it's error status correctly, only
telling you that the error was present if another component on the same page also was displaying an
error. Changing it to a string fixes the issue, and matches how we are getting patterns from the server
so that's it for now.