Workaround for https://github.com/gobuffalo/pop/issues/72
This above issue made it so that if Eager was called on a
*pop.Connection it made all future queries Eager. This broke us because
it meant that some of our Creates() became Eager when we didn't expect
them to. (it crashed because they don't expect us to use pointers)

This fix copies the pop.Connection before creating a query so that the
Eager call doesn't poision the global connection.