Deploy app to a Fargate service
This migrates the app to be deployed on ECS Fargate (instead of EC2).

As part of the migration, we remove the cpu and memory values as they
only make sense for EC2 containers. In Fargate, the cpu and memory are
defined as part of the task definition, not the container definition.