Entering a Move Date updates default calculator dates (#1805)
* Resolve an issue where entering a departure date wouldn't cause the Move Date
fields in the calculators to be updated.

The cause of the issue is that we were dynamically naming a <Field>, which isn't handled
by redux-form by default. Setting a new name for the field caused it to not be considered
"registered" by that library, which prevented the field's value from being updated.

In this case, though, there wasn't a reason for the calculators to be aware of how the move
date they used was determined. By moving the derivation of the move date into mapStateToProps,
we can avoid the behavior of redux-form described above and have the updated move date
propagated as we would expect.

* Fix broken e2e tests.

* Update SIT calculator to use move_date.