Don't run client build while starting server_run
We were doing this to create a file at build/index.html that the
milmove serve command needs to read in when it starts up. However,
this index file is not requested in local development as the index
is served by the create-react-app development server.

This commit puts an empty file in that location so that we don't need
to run client_build just to start up the server. In production we don't
invoke the server through make, so this will have no effect in those
environments.