Add last_sign_in_at and last_sign_out_at to users
**Why**: One of the STIG requirements is to log the last time a user
signed in and out. Adding these 2 fields prepares us to implement that
feature.

I don't believe the STIG specifically requires that this be tracked in
the DB, but from past experience, using a DB is more useful than
logging to CloudWatch, for example, because it makes it easier to
run queries.