Add Ruby script to automate prod deploys
**Why**:
The current deploy process followed twice a day by the folks who are
on call is manual and tedious. It requires following these steps:

1. Visit the Circle CI website to view the most recent pipelines for
the master branch
2. Click on the most recent successful pipeline
3. Copy the git commit hash
4. Look up the date for the commit hash via the command line
5. Visit the GitHub website to view pull requests for mymove
6. Filter PRs using this query:
`is:closed base:master merged:>2020-05-04T17:33:46-00:00 sort:updated-asc`
where the date is the one from step 4, but with one second added to
it, and converted to ISO 8601 format that GitHub needs.These first 6
steps are how we can obtain all PRs that have been merged to master
since the last one that was deployed to production.
7. Copy the title of each PR into a new Slack post and share it in
the #prod-support channel
8. Go back to the CircleCI pipelines, and click on the most recent
"ON_HOLD" pipeline.
9. Scroll horizontally until you see the "approve_prod_deploy" job,
and click on it, then type in "approve_prod_deploy" to approve it.

These steps are time-consuming and error-prone. With this script, the
process is reduced from several minutes to a few seconds. If there
aren't any new merged PRs since the last deploy, it will let you know.
If there are PRs to deploy, it will first prompt you to make sure you
really want to deploy. This helps avoid accidental deploys.

For troubleshooting and comparison with the CircleCI website, the
script will output the pipeline number for the most recent successful
and on hold pipelines, and the commit hash for the last deployed PR.
For easy copy and pasting into Slack, it will output the title and URL
for each PR that is about to be deployed.