MB-2329 Revoke sessions via admin interface
One of the STIG requirements is the ability to revoke individual
sessions, and because we don't have access to the Redis server in
prod, we need to provide a way for someone to do this via the admin
website.

**Setup**:
`make db_dev_migrate`
`make redis_pull`

**To test session revocation:**

1. Sign in at http://milmovelocal:3000/ in one browser tab
2. In another browser tab, sign in at http://adminlocal:3000/ (You
might need to create yourself as an admin first via
http://adminlocal:3000/devlocal-auth/login)
3. Once in the admin site, click on Search User by ID on the left
sidebar
4. Enter your UUID and click the Search button. You can find your UUID
in the server logs in the middleware/request_logger.go:85 line under
the key `user-id`.
5. Once your details appear in the user show page, click on Edit in
the top right
6. Select "Yes" from the dropdown for "Revoke mil session"
7. Go back to the tab where you were signed into the mil app and
refresh it.
8.Verify that you are logged out, and that the session token has been
removed from the `current_mil_session_id` field in the Users table,
and from Redis (accessed via `redis-dev`, then list all keys with
`KEYS *`)