Fix checks around session config settings
When checking if the session idle timeout and lifetime are within our
limits, we were not comparing the same time units. Furthermore, this
check was only happening in prod, so we missed this bug in staging.

This fixes the checks so that they compare the same time units, and
allows the checks to run in all deployed environments.