[MB-3061] Consolidate service_members & customers
This is an initial minimal pass at merging the customers table into
service_members.

There is almost 100% duplication between the service_members table and
the customers table at the moment. The new onboarding process does not
seem to suggest any reason why these should need to remain separated.
The customer profile data collected appears to be the same regardless of
the type of move you're making. So, this PR consolidates that data into
a single table going forward.

To minimize the breadth of the changes (and risk of breaking
production), this first pass attempts to migrate the data and change
just the database interactions on the customers side to make them use
the service_members table instead. No APIs or payloads have been changed
yet. Separate stories will handle API rework and potentially changing
the service_members naming to customers.