MB-3113 Consolidate move_orders and orders tables
**Description**
This is part of the ongoing effort to reduce duplication between
certain tables, which is causing confusion and pain in implementing
some features.

Because the move_orders table is not used in production, we opted to
consolidate move_orders into orders. In staging and experimental, the
move_orders (and related tables) were populated, and because the
orders table requires a document to be associated with an order, if we
wanted to to preserve the data in staging, that would mean creating
a document for each move_order. In addition, the orders_type column
in move_orders was of type enum, versus varchar in the orders table.
Rather than spend time preserving and converting data, we opted to
delete all the data in staging by using the Postgres TRUNCATE and
CASCADE commands, which clears out the following tables:
move_orders, move_task_orders, mto_service_items, payment_requests,
mto_shipments, payment_service_items, mto_service_item_dimensions,
mto_service_item_customer_contacts, proof_of_service_docs, mto_agents,
and payment_service_item_params.

For the difference in the orders_type column type, we opted to leave
the orders column as a varchar for now, and add the enum values
defined in the move_orders table ("GHC" and "NTS") into the enum
values defined for the orders table in the swagger `internal.yaml`
config. The reasoning behind that is that converting a varchar column
into an enum requires 5 separate deployments for a zero-downtime
migration. We thought it was better to get the tables consolidated
first with the least amount of changes, and then go back and make
changes if desired.

Similarly, when adding columns into orders that previously only
existed in the move_orders table, we preserved the type, even though
it resulted in inconsistencies within the orders table. For example,
the move_orders table used `*DutyStation` for the `OriginDutyStation`
whereas the orders table uses `DutyStation` for the `NewDutyStation`.

Note that this PR affects the Support API:

- When creating a move task order, the associated move order is
expected to be associated with an existing document, so the document
must be created first (tied to the same service member). Other fields
that were required by the orders model and that are now also required
are: status, issue date, and orders type.

The GHC API was also changed to account for the different type used for
OrdersType and OrdersTypeDetail. In orders, the type is
`internalmessages.OrdersType`, versus String in move_orders.