Remove references to customer destination address
**Why**:
In January, we added a new destination_address field to the customers
table based on the thinking that this might be something the service
member would fill out during onboarding. Here is the original PR and
JIRA story:
https://github.com/transcom/mymove/pull/3291/files
https://dp3.atlassian.net/browse/MB-1051

Since the introduction of that field, it has not been used in the code
other than via the testdatagen tool solely for the purpose of testing
that if the field is populated, it shows up in the front end.

When we consolidated the customers and service members tables, we
removed the destination address field because we saw it was not being
used, but we missed the references to it in the front end.

Looking at the latest onboarding design mocks, there is no reference
to a destination address that would be part of the service member's
profile. Only shipments have a destination address.