MB-3474 Remove references to move order in tests (#4504)
* MB-3474 Remove references to move order in tests

This is some small follow-up clean up work after consolidating the 
`orders` and `move_orders` tables. In the previous PR, we limited the
changes to only those necessary to merge the tables and nothing more,
to keep things simple and to avoid breaking APIs. This left references
to "move orders" even though the table and model are now gone.

In this PR, to keep things simple and safe, we are only removing
references to move orders in test data generation files and related
tests. We will clean up the rest of the app in future PRs.

**Reviewer Notes**

Previously, the test data generation for "move orders" created service 
members with random names and grades. That was not the case with the 
test data generation for orders, so we are no longer generating random 
data here unless a field is meant to be unique. We don't seem to 
generate random data in many other places in testdatagen, so I'm 
wondering what we may be losing by no longer doing this.

I would like a review from folks who depend on local data being available
in the TOO interface to make sure all the expected data is still being generated.

**Setup**
```
make clean && make build
make db_dev_reset && make db_dev_migrate
generate-test-data --named-scenario e2e_basic
```