MB-3504 Remove GHC and NTS from OrdersType
When we consolidated the `orders` and `move_orders` tables, we noticed
that the `move_orders` table had an enum column called `OrderType`
with two possible values: GHC and NTS. To preserve this existing
functionality, we added "GHC" and "NTS" to the list of enums defined
under `OrdersType` in the `internal.yaml` Swagger file.

It has now come to our attention that "GHC" and "NTS" are not valid
options for `OrdersType` and should be removed.