MB-3114 Consolidate MoveTaskOrders with Moves
Things to note:
In addition to the fields we moved over from MTOs, I also had to move
over the associations. See the changes in the Move model.
To deal with `is_canceled` not being ported over, I kept the reference
to `IsCanceled` in the APIs and added a new function to return the
correct boolean depending on the `status` of the move. This function
is duplicated in several places. Is it worth it to extract it? If so,
what's the best place for that?    API changes: Show, Locator, Status
are now required in support  Debugging notes: column move_id does not
exist in move_task_order_fetcher.go
```
err := f.db.Where("moves.orders_id = $1", moveOrderID).Eager().
  All(&moveTaskOrders)
```
That's because of Pop Eager loading associations like MTOshipments
that still refer to `move_task_order_id` but Pop expects it to refer
to `move_id`. That was the same issue with query builder. So, I had to
revert to the old migration that renames the foreign key columns.

why does pop complain that `show` can't be null if it has a default
value in PG? is it because it's a pointer? For example, in
`move_task_order_creator_test.go`

where we have tests for reference id, make sure we have one for
locator as well.

In `handlers/supportapi/move_task_order_test.go`, we were setting the
Reference ID to 346523, which doesn't match the format xxxx-xxxx.
Sounds like we don't have a validation for the format.