MB-3569 Treat HHG and HHG_LONGHAUL_DOM the same
**Why**:
Currently, when a customer goes through onboarding and selects HHG, we
create an MTO Shipment with a shipment_type of `HHG`. However, in our
`mto_shipment_updater`, which is what creates service items when the
TOO approves the shipment, we don't handle the case where the shipment
type is just plain `HHG`.

This is blocking other teams, namely John Gedeon, from testing their
work that relies on service items being created. To work around that,
they have created a `convert.go` file that adds complexity and looks
up current moves that were created via onboarding, and adds longhaul
and shorthaul shipments to it, simply to be able to approve them via
the TOO interface and therefore have service items created.

The `convert.go` file has been a source of confusion during our DB
consolidation efforts, and we can get rid of it by duplicating the
current logic in the switch statement in `mto_shipment_updater` for
both `HHG` and `HHG_LONGHAUL_DOMESTIC` shipment types until we
decide whether or not there is a different between `HHG` and other
shipment types, and if we should default to `HHG_LONGHAUL_DOMESTIC`
instead of `HHG`.