MB-3591 Fix MTO shipment updater tests
These 2 integration tests that tested updating a shipment's status to
`approved` were using a shipment of type `HHG`, which isn't handled by
the code that performs the update in `mto_shipment_updater`. These
tests were bypassing the code they were attempting to test. In order
to test the updater, one of the supported shipment types must be used,
such as `HHG_LONGHAUL_DOM`. In turn, the ReServices table must be
populated with the appropriate codes for the shipment to be updated.