Remove invalid Cypress test and increase timeout
We are currently seeing failures in Circle CI on the master branch for
tests in `ppmCloseout.js`, which has happened several times in the
past. The failure has always been due to the test not finding the
element in time. By default, Cypress will only wait up to 4 seconds
total for a command and its assertions to execute.

To increase the timeout, we can pass in the timeout option to a
command, which is what I've done here for the most common failures.
I've increased it to 30 seconds to see if that helps.

I also removed a test that was looking for the word "unknown" in the
`ApprovedMoveSummary.jsx`. My guess is that the "unknown" was
referring to the estimated weight, which is no longer unknown,
possibly due to a recent change, or this one back in May:
https://github.com/transcom/mymove/pull/4088