MB-3114 Consolidate moves and MoveTaskOrders (#4533)
* Added migration to consolidate move_task_orders into moves

* Removing move_task_order_id name change for now

* MB-3114 Consolidate MoveTaskOrders with Moves

Things to note:
In addition to the fields we moved over from MTOs, I also had to move
over the associations. See the changes in the Move model.
To deal with `is_canceled` not being ported over, I kept the reference
to `IsCanceled` in the APIs and added a new function to return the
correct boolean depending on the `status` of the move. This function
is duplicated in several places. Is it worth it to extract it? If so,
what's the best place for that?    API changes: Show, Locator, Status
are now required in support  Debugging notes: column move_id does not
exist in move_task_order_fetcher.go
```
err := f.db.Where("moves.orders_id = $1", moveOrderID).Eager().
  All(&moveTaskOrders)
```
That's because of Pop Eager loading associations like MTOshipments
that still refer to `move_task_order_id` but Pop expects it to refer
to `move_id`. That was the same issue with query builder. So, I had to
revert to the old migration that renames the foreign key columns.

why does pop complain that `show` can't be null if it has a default
value in PG? is it because it's a pointer? For example, in
`move_task_order_creator_test.go`

where we have tests for reference id, make sure we have one for
locator as well.

In `handlers/supportapi/move_task_order_test.go`, we were setting the
Reference ID to 346523, which doesn't match the format xxxx-xxxx.
Sounds like we don't have a validation for the format.

* Remove duplicated reference ID generation

* Clean up

* Hardcode show to true

* Locator is created when not in MTO payload

* Address PR feedback comments

* Default to draft status

* Add tests for IsCanceled

* Use pointers for referenceID and contractorID

* Remove convert.go

* Fix migration

* Limit orders to those whose moves have shipments

* Use 2 inner joins

* Update migration after merging with master

* Deploy to experimental

* Re-mapped new references to move_task_order table from merge

* Address John's comments

* Reset Circle config to deploy to staging

* Revert "Remove convert.go"

This reverts commit 9fb3d680ef941c77b5a8e5695bba5b7bb2db80e4.

* Add conversion function back in

* Deploy to experimental (again)

* Revert "Deploy to experimental (again)"

This reverts commit 2ca3c08ede3a67bfa32b20c8441af60f9d43bfee.

Co-authored-by: Reggie Riser <reggie@truss.works>