MB-3540 Remove references to MTOs in tests
Now that the Moves/MTOs consolidation is done, we can start some of
the cleanup work. In this PR, we are replacing references to
`MoveTaskOrder` with `Move`, but only in the testdatagen functions and
in tests that call those functions. Note that there are still
references to `MoveTaskOrder` within models. Those will be cleaned up
separately.