MB-3556 Keep testdatagen functions organized
During our work to consolidate Moves and MTOs, we noticed this
`MakeAvailableMoveTaskOrder` function that fit better inside the
`testdatagen` package. This PR moves that function into the
`make_move.go` file and renames it to `MakeAvailableMove` since we
no longer have a Move Task Orders table.

In addition, we replaced instances of creating a Move with
`AvailableToPrimeAt` set to the current time with a call to
`MakeAvailableMove`.