MB-3667 Create order with all fields at same time
During onboarding, we store the service member's rank and origin duty
station on the `service_members` table. However, on the TOO page,
those fields are assumed to come from the `Orders` table. Since those
fields are expected to be on the `Orders` table, they should be
populated at the same time as the order is first created. This was not
the case. The way those fields were added to the Order was via a
temporary hack.

During the moves and MTOs consolidation, we were able to remove one of
the temporary conversion functions. In this PR, we are able to
remove the remaining conversion-related code entirely by adding those
missing order fields to the order creation in the orders handler.