MB-3662 Add comments to tables assigned to TTV
This is part of the ongoing effort to have our entire database
documented by the end of August. Each team has been assigned a set
of tables.

Note that some tables and columns listed in the JIRA story are not
documented here because they are no longer used. I will submit a
separate PR to remove unused tables and related code.