MB-3712 Remove unused role-based auth code
The code removed in this PR was written 8 months ago by Team Roci (
Matt Krump and Ren specifically), but the only way this code would
ever get executed is if the feature flag `FEATURE_FLAG_ROLE_BASED_AUTH`
was set to `true`. This flag defaulted to `false`, and is currently
not set in any environment, and has never been as far as I can tell.

The idea behind this code was to protect API endpoints so that only
users with the appropriate roles could access them. The implementation
was based on [ADR #0040](docs/adr/0040-role-base-authorization.md). A
new "x-swagger-roles" tag was added to a couple of the Swagger YAML
files on each endpoint, indicating the list of roles that can access
the endpoint. The `RoleAuthMiddleware` was added to compare this list
with the roles the user has (by checking the session).

Here are the 3 original PRs that implemented this middleware-based
authorization:

https://github.com/transcom/mymove/pull/3177
https://github.com/transcom/mymove/pull/3182
https://github.com/transcom/mymove/pull/3217

Why this was never turned on, nobody seems to know. Since then, we
implemented something similar, but on the front end only, and only
for the office app. We check a user's role before giving access to
an office endpoint. Here is the original PR:
https://github.com/transcom/mymove/pull/4069

This code seems useful to me, so I'm surprised it was never used.
We should find out whether we do need these protections on all the
endpoints that were originally meant to be protected, and if we still
want to follow ADR 0040. If so, I would still merge this and open a
new story to rewrite the authentication so that it uses this
role-based middleware.