MB-3550 Show backup contact info in Move details
In order to display backup contact info on the Move details page (as
seen by a TOO), the GHC API had to be modified to include that data.

Note that the Service Member model refers to Backup Contacts as a
`has_many` relationship, but in practice, a service member can only
add a single backup contact during onboarding. For that reason, I
named the key in the payload `backup_contact`.

For now, I only included the name, email, and phone attributes in the
swagger definition. If it's expected that a TOO should be able to
make changes to the backup contact, then we'll need to add the `id`
and `eTag`.

Because the backup contact `phone` is optional, I made it appear last
in the Customer info component.