MB-4299 Encrypt S3 uploads
**Description**
The move to GovCloud has enabled a service control policy (SCP) to
reject unencrypted uploads to S3. Thankfully, the `aws-sdk-go` package
makes it easy for the client to specify that uploads should be
encrypted.

Since both user and Prime uploads use the same S3 Storer object, we
only need to make a one-line change. As for secure migrations, they
were already using the `--sse AES256` flag.

**Reviewer Notes**
This PR is currently deployed in experimental. Please test the
following:
- Upload a document as a service member during onboarding
- Upload a proof of service via the Prime API
- Upload a secure migration

Note that we currently cannot visually confirm the user upload in the
office app (as a TOO) due to a known CORS issue that is being fixed
this week.