MB-4384 Display empty string for 0 SIT allowance
By default, a service member's SIT (storage in transit) allowance is
90 days. However, the `StorageInTransit` field does not have a default
value in the DB, and we are not explicitly setting it to 90 days
when the customer creates their move. This was causing the frontend
to display `undefined days` as the SIT Allowance in the Move Details
page in the TOO interface. This is because the `StorageInTransit`
field was not being returned at all as part of the payload when it is
zero. Note that there is a separate JIRA story to set the default
SIT allowance.

In this PR, we are fixing the frontend so that it displays an empty
string if SIT is zero. To do that, we set the `x-nullable` attribute
for `storageInTransit` to `true`, which allows the frontend to check
if it is zero or non-zero. If it is zero, we display an empty string.

If it's ever possible for the SIT allowance to only be one day, we
might consider using some kind of pluralization helper so that it
would display the singular `day` as opposed to the hardcoded plural
`days` that we have now.