Extract eslint and pretter from pre-commit in circle ci
The way pre-commit is configured to run eslint and prettier was
resulting in many many processes being started each with a small portion
of the javascript files we have. This was overloading the resources on
the circle ci container. This commit skips eslint and prettier in the
pre-commit step and adds 2 steps to run eslint and prettier on their own
in more standard ways that use much less resources