MB-4385 Allow MTO tab to render all the time
**Description**
Previously, if you clicked on the Move task order tab when looking
at a Move as a TOO, and when no shipments had been approved yet,
it would throw an exception. This was because the code was assuming
there would always be service items present.

To fix this, we can use the Null Propagation operator so that if
something is null, it doesn't throw an `undefined` exception. In
addition, we added a conditional around the ShipmentContainer so that
if no approved shipments exists, we display an appropriate message,
and if there are multiple shipments, of which only some are approved,
we only display the approved ones.