MB-4489 Show ZIP code if shipment address missing
**Description**
If the service member doesn't know their full destination address for
the shipment, the form defaults to their new duty station ZIP code.
On the Move task order tab in the Office app, we were wrongly assuming
every shipment would have a destination address, resulting in the
string `undefined` for the city, state, and ZIP.

To fix this, we check to see if a destination address exists, and if
not, we fall back to the postal code of the order's new duty station.

I also fixed a formatting issue where a comma was missing between the
city and state of the pickup address.