Use pre-commit-golang to run golangci-lint
When the Go version is managed by `asdf`, the golanci-lint pre-commit
hook fails to find it. I believe the issue is how we're calling it
in the `entry` section. I found a repo that compiles a bunch of Golang
hooks for use with pre-commit, and the way they configure it works
with `asdf`.

Note that this requires that everyone install golangci-lint with
Homebrew, which is why I added it to `scripts/prereqs`.