MB-4061 Filter payment requests by GBLOC (#4959)
* MB-4061 Filter payment requests by GBLOC

**Description**
When a TIO signs in, they should only see the payment
requests associated with a transportation office that has the same
GBLOC as the transporation office the office user belongs to.

This PR also adds authorization to the list payment requests endpoint
so that it only allows users who have the TIO role.

In order to speed up tests and avoid using the DB when not needed, I
added a new top-level `Stub` assertion type to `testdatagen` so that
you can bypass the DB when setting it to `true`. The testdatagen
functions are very convenient, but sometimes we just want plain models
without saving them to the DB. By setting the `Stub` assertion, we can
take advantage of that convenience without having to write the models
from scratch in tests.