Allow finding matching duty stations and TOs
In the October slice, TTV is working on filtering the TOO and TIO
queues based on the GBLOC. When we demo this work, we have to make
sure that the Origin Duty Station chosen by the service member
corresponds to a transportation office with the same GBLOC as the
one the TOO and TIO are assigned to.

Similarly, to demo the scenario where the GBLOCs don't match, we need
to choose a transportation office for the TOO and TIO that is in a
different GBLOC than the one associated with the duty station.

To make it easier to find the names of matching transportation
offices or duty stations, I've created these two scripts.

The `ds_for_gbloc` script expects a transportation office name, and it
will show you all the duty stations with a matching GBLOC.

The `to_for_gbloc` script expects a duty station name, and it
will show you all the transportation offices with a matching GBLOC.

A usage example during demo prep would be:

1. A TOO/TIO already exists, and is assigned to transportation office
X
2. The person playing the role of the service member will run
`ds_for_gbloc <name of transportation office from step 1>`, and will
pick any of the results as their origin duty station

Alternatively:

1. The person playing the role of the service member will agree to
pick duty station X
2. The person playing the role of the TOO/TIO will run
`to_for_gbloc <name of duty station from step 1>` and update their
transportation office (via the admin interface) to any office in the
script results.