Speed up tests
This PR includes various ways to speed up our tests. I tried to
break them down into small chunks in seperate commits to make it
easier to review.

The first change was to create a `MakeStubbedUser` function in
testdatagen that uses the new `Stub` assertion (which was added in
a previous PR) so that it doesn't touch the database. I then replaced
all instances of `MakeDefaultUser` with `MakeStubbedUser` where we
don't actually need a user in the DB.

The second change was to add the `-failfast` and `-vet=off` flags to
our `go test` command so that we don't have to wait for the entire
test suite to finish just to tell us that a test failed. Note that the
failfast flag is only enabled locally. In Circle CI, we want to know
about all test failures at once.