Specify node version in pre-commit config (#5035)
* Specify node version in pre-commit config

**Why**: By default, pre-commit expects a language to be available
globally. We use nodenv to manage node versions, and we shouldn't
expect folks to set their global version to always match the one we
use on the project.

To tell pre-commit to use a specific version of a language, you can
use the `language_version` key or `default_language_version`. 
This works for `node`, but not for `golang` unfortunately. 
When I tried it, I got this message:

```
An unexpected error has occurred: AssertionError: For now, pre-commit
requires system-installed golang
```

I verified that when I didn't have `node` installed globally, the
`pre-commit install-hooks` command hung when trying to install the
node-based hooks. After I added the `default_language_version`,
everything installed as expected without hanging.