MB-4845 TOO queue can filter status (#5077)
As a TOO, I want to be able to filter certain columns in my queue, so that I can narrow my focus on relevant moves.
> Given that I am a TOO logged into MM
>  When I look at my work queue and want to view a subset of data
>  Then I can use filters on the following columns:
>       Status

There are a few different types of filters for the various columns:

MoveStatus - Multi drop down selector, with options:

- New move

- Approvals requested

- Move approved

This PR only implements the backend portion of the filter. The multi-select dropdown and ability to query from the frontend will be implemented in a separate JIRA story.