MB-4997 Remove old order & payment request queues
To reduce duplication, we would like to define the query filters
(that are currently defined in the `queues.go` handler) in a service.
That way, we can reuse them across handlers, which we will need to do
for the payment requests queue.

To move the query filters from the handler to a service, the handler
has to pass the `params` to the service. The problem is that the old
`ListPaymentRequestsHandler` and the new
`GetPaymentRequestsQueueHandler` use different types for their params,
but they both call the same `FetchPaymentRequestList` service. The
same is true for the old and new handlers that fetch orders.

Since we no longer need the old handlers, we can remove them, and
update the frontend office index page to automatically route to the
new queue endpoints.