Add new script to seed dev DB
**Why**: We currently use the `e2ebasic.go` file for seeding the
test DB for Cypress end-to-end tests, as well as for seeding the dev
DB for testing locally. The problem is that as we add more things to
create in the DB that are only for local dev use, it increases the
time it takes to populate the test DB before the Cypress tests can
run. There's also the risk that someone might change data not
realizing that it can affect the end-to-end tests.

It's better to separate the two so that they don't affect one another.
To that end, I created a new `devseed.go` file that is currently a
duplicate of the `e2ebasic.go` file. This will allow us to stop using
the `e2ebasic.go` file for things that are not related to Cypress
tests.