Refactor devseed for easier batch creation
By creating functions for each type of entry, it makes the file more
readable, and makes it a lot easier to create many instances of a
particular entry. For example, to test pagination in the TXO queues,
you can now easily loop over one of the functions.

I removed some entries that didn't seem to be used for local testing.

I also noticed that in some cases, we are using the same email for a
bunch of users, which doesn't reflect reality. Similarly, we are using
the default ARMY affiliation for all service members. Ideally, we would
make email unique and choose a random affiliation, which will make it
easier to test sorting and filtering in the TXO queues. I will do that
in a separate PR.