MB-5457 Fix sort order of moves in TOO queue (#5195)
* MB-5457 Fix sort order of moves in TOO queue

We want to sort by the move status, but we were sorting by the order
status. The easiest way to fix this is to have the query return moves
instead of orders.

The payload is returning a list of moves, and the UI also refers to
the list as a collection of moves. To avoid confusion, the DB query
should return moves as well.