Fix display of status in TOO queue
In PR #5158, we made a change to add a new `APPROVALS REQUESTED`
status enum, which simplified the way we determine a move's status,
and allowed us to implement pagination.

Previously, the API would perform the translation from the value
stored in the DB to the TOO-friendly one we want to display. For
example, `SUBMITTED` became `New move`. After the change in #5158, the
payload was no longer translating and was returning the DB value.

This commit now performs the translation on the frontend.