MB-4829 Include dependents in weight allowance (#5331)
* MB-4829 Include dependents in weight allowance

Previously, the GHC API payload, when returning an order, was setting
the `totalWeight` key to the weight allowance based only on the
service member's rank, regardless of whether or not they have
dependents.

This PR fixes the bug in the simplest way as a first pass, by adding
a conditional to check if the entitlement includes dependents, and if
so, it uses the `TotalWeightSelfPlusDependents` key in the lookup
table in `entitlements.go`. Otherwise, it uses `TotalWeightSelfPlus`
as before.

I added 2 handler tests for both scenarios.