Attempt to fix Cypress flakes
For some reason, we are making two requests on the TXO pages: one with
page, per page, and order, and then a second request with the same
params, but this time with sort added to the mix.

Because we were using a glob to match all requests to the TXO
endpoints, calling cy.wait will stop at the first match, and will not
also wait for the second request. I added a second named route to
match the second request, and we are now waiting for that one as well.